+++ raiden/tests/e2e/raiden.spec.ts
@@ -15,7 +15,8 @@ import { request } from 'matrix-js-sdk';
 import { Raiden } from 'raiden/raiden';
 import { ShutdownReason } from 'raiden/constants';
 import { initialState } from 'raiden/store';
-import { raidenShutdown, newBlock } from 'raiden/store/actions';
+import { raidenShutdown } from 'raiden/store/actions';
+import { newBlock } from 'raiden/channels/actions';
 import { ChannelState } from 'raiden/channels';
 import { Storage } from 'raiden/utils/types';
 import { ContractsInfo, RaidenContracts } from 'raiden/types';
